+++ php_phongo_structs-5.h
@@ -120,7 +120,7 @@ typedef struct {
 	zend_object              std;
 	char                    *data;
 	int                      data_len;
-	int                      type;
+	uint8_t                  type;
 } php_phongo_binary_t;
 typedef struct {
 	zend_object              std;
